Created Trondheim folder
