+++ src/pages/public/Team/TeamPages/index.ts
@@ -1,8 +1,8 @@
-export { default as IT } from "./IT";
-export { default as Evaluering } from "./Evaluering";
-export { default as Okonomi } from "./Okonomi";
-export { default as Profilering } from "./Profilering";
-export { default as Rekruttering } from "./Rekruttering";
-export { default as Skolekoordinering } from "./Skolekoordinering";
-export { default as Sponsor } from "./Sponsor";
-export { default as Styret } from "./Styret";
+export { default as IT } from "./Trondheim/IT";
+export { default as Evaluering } from "./Trondheim/Evaluering";
+export { default as Okonomi } from "./Trondheim/Okonomi";
+export { default as Profilering } from "./Trondheim/Profilering";
+export { default as Rekruttering } from "./Trondheim/Rekruttering";
+export { default as Skolekoordinering } from "./Trondheim/Skolekoordinering";
+export { default as Sponsor } from "./Trondheim/Sponsor";
+export { default as Styret } from "./Trondheim/Styret";
